Adding license to let people know the rules
This databse is creative commons non-commericial. So commercial developers need to get permission, but open source developers working on projects just go for it. The point of adding it here was to increase the access and get more people using this datasets in their applications or being inspired by this dataset to write software.